// src/types/supabase.ts
export type Json =
	| string
	| number
	| boolean
	| null
	| { [key: string]: Json | undefined }
	| Json[]

export type Database = {
	public: {
		Tables: {
// ...
			matches: {
				Row: {
					created_at: string
					date: string
					forfeited_team: string | null
					id: number
					match_id: number
					referee: string | null
					round: string
					team1_id: number
					team1_score: number
					team2_id: number
					team2_score: number
					time: string
					type: Database['public']['Enums']['match_type']
				}
				Insert: {
					created_at?: string
					date: string
					forfeited_team?: string | null
					id?: number
					match_id: number
					referee?: string | null
					round: string
					team1_id: number
					team1_score: number
					team2_id: number
					team2_score: number
					time: string
					type: Database['public']['Enums']['match_type']
				}
				Update: {
					created_at?: string
					date?: string
					forfeited_team?: string | null
					id?: number
					match_id?: number
					referee?: string | null
					round?: string
					team1_id?: number
					team1_score?: number
					team2_id?: number
					team2_score?: number
					time?: string
					type?: Database['public']['Enums']['match_type']
				}
				Relationships: [
					{
						foreignKeyName: 'public_matches_team1_id_fkey'
						columns: ['team1_id']
						isOneToOne: false
						referencedRelation: 'teams'
						referencedColumns: ['id']
					},
					{
						foreignKeyName: 'public_matches_team2_id_fkey'
						columns: ['team2_id']
						isOneToOne: false
						referencedRelation: 'teams'
						referencedColumns: ['id']
					}
				]
			}
			players: {
				Row: {
					created_at: string
					joined_at: string
					role: Database['public']['Enums']['player_role']
					team_id: number
					updated_at: string
					user_id: string
				}
				Insert: {
					created_at?: string
					joined_at: string
					role: Database['public']['Enums']['player_role']
					team_id: number
					updated_at: string
					user_id: string
				}
				Update: {
					created_at?: string
					joined_at?: string
					role?: Database['public']['Enums']['player_role']
					team_id?: number
					updated_at?: string
					user_id?: string
				}
				Relationships: [
					{
						foreignKeyName: 'players_team_id_fkey'
						columns: ['team_id']
						isOneToOne: false
						referencedRelation: 'teams'
						referencedColumns: ['id']
					},
					{
						foreignKeyName: 'players_user_id_fkey'
						columns: ['user_id']
						isOneToOne: true
						referencedRelation: 'users'
						referencedColumns: ['osu_id']
					}
				]
			}
			teams: {
				Row: {
					acronym: string
					availability: Json | null
					available_ending: string | null
					available_starting: string | null
					created_at: string
					flag: string
					id: number
					name: string
					qual_lobby_id: number | null
					timezone: string
					updated_at: string
				}
				Insert: {
					acronym: string
					availability?: Json | null
					available_ending?: string | null
					available_starting?: string | null
					created_at?: string
					flag: string
					id?: number
					name: string
					qual_lobby_id?: number | null
					timezone: string
					updated_at: string
				}
				Update: {
					acronym?: string
					availability?: Json | null
					available_ending?: string | null
					available_starting?: string | null
					created_at?: string
					flag?: string
					id?: number
					name?: string
					qual_lobby_id?: number | null
					timezone?: string
					updated_at?: string
				}
				Relationships: [
					{
						foreignKeyName: 'public_teams_qual_lobby_id_fkey'
						columns: ['qual_lobby_id']
						isOneToOne: false
						referencedRelation: 'lobbies'
						referencedColumns: ['id']
					}
				]
			}
// ...
			users: {
				Row: {
					country: string
					country_code: string
					country_rank: number | null
					created_at: string
					discord_avatar: string | null
					discord_id: string
					discord_name: string | null
					discord_tag: string
					osu_avatar: string
					osu_id: string
					osu_name: string
					rank: number | null
					restricted: boolean
				}
				Insert: {
					country: string
					country_code: string
					country_rank?: number | null
					created_at?: string
					discord_avatar?: string | null
					discord_id: string
					discord_name?: string | null
					discord_tag: string
					osu_avatar: string
					osu_id: string
					osu_name: string
					rank?: number | null
					restricted: boolean
				}
				Update: {
					country?: string
					country_code?: string
					country_rank?: number | null
					created_at?: string
					discord_avatar?: string | null
					discord_id?: string
					discord_name?: string | null
					discord_tag?: string
					osu_avatar?: string
					osu_id?: string
					osu_name?: string
					rank?: number | null
					restricted?: boolean
				}
				Relationships: []
			}
		}
// ...
	}
}

type PublicSchema = Database[Extract<keyof Database, 'public'>]

export type Tables<
	PublicTableNameOrOptions extends
		| keyof (PublicSchema['Tables'] & PublicSchema['Views'])
		| { schema: keyof Database },
	TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
		? keyof (Database[PublicTableNameOrOptions['schema']]['Tables'] &
				Database[PublicTableNameOrOptions['schema']]['Views'])
		: never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
	? (Database[PublicTableNameOrOptions['schema']]['Tables'] &
			Database[PublicTableNameOrOptions['schema']]['Views'])[TableName] extends {
			Row: infer R
		}
		? R
		: never
	: PublicTableNameOrOptions extends keyof (PublicSchema['Tables'] &
				PublicSchema['Views'])
		? (PublicSchema['Tables'] &
				PublicSchema['Views'])[PublicTableNameOrOptions] extends {
				Row: infer R
			}
			? R
			: never
		: never
// ...
export type Enums<
	PublicEnumNameOrOptions extends
		| keyof PublicSchema['Enums']
		| { schema: keyof Database },
	EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
		? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
		: never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
	? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
	: PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
		? PublicSchema['Enums'][PublicEnumNameOrOptions]
		: never